Test against an expected result
This exposes an issue: our expectations do not hold, because Array and
Item are mutated in place. A solution for that would be to make sure
that we start working with identical arrays that are, however, not the
same arrays.

Careful, this loses a property: in the public API, the items in the
array  must be the same references!