Implement regression tests as a constant gen
Our test looks good, but the generation is not great yet - we're
always generating the same element. Let's try to generate truly random
items.

Note that we'll not respect legal values, for the simple reason that
we want the legacy and modern implementations to behave the same for
legal *and* illegal scenarios.