Add an Item extractor
The purpose is to make the quality update more idiomatic. Once I'm
done, it should look a lot more like case classes being
pattern-matched on.

This is slightly dodgy - we were instructed not to touch Item.

I feel this respects the letter of the law - I'm not touching the Item
class, but its companion object - if not the word. If this is
considered unnacceptable, the extractor can always be moved to another namespace.